+++ PS.Mothership.Core/PS.Mothership.Core.Common/Dto/Merchant/OfferAuthorizationFeeDto.cs
@@ -2,7 +2,7 @@ namespace PS.Mothership.Core.Common.Dto.Merchant
 {
     public class OfferAuthorizationFeeDto
     {
-        public long AuthorisationFeeKey { get; set; }
+        public int AuthorisationFeeKey { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public decimal Value { get; set; }
